V1.1 Update - Oval Topper Mod Added
The ball topper version of the joystick was out of stock for a while so we created an alternate version
that works with the oval topper version.
The enclosure is the same but the thread for the topper is a M10 not an M6. Therefore, a new topper
nut was designed. This was just created to have an insert for a M10 nut without the catch.

Documentation was updated as well.